Fix Issue #18
At the SpinningCursor#stop method the 'release_console'
was being called while the "@spinner" was still spinning.
So that the @spinner tries to call $stdout.string after console released
and there's no 'string' method for the STDOUT (as $stdout points to it when
"released")

Inverting the order and killing the spinning before releasing console
fix the issue.